Save and move states

+++ TelegramAnki/Anki.DAL/DTOs/UserDTO.cs
@@ -3,7 +3,7 @@ namespace Anki.DAL.DTOs
 
     public class UserDTO
     {
-        public int ChatID { get; set; }
+        public long ChatID { get; set; }
         public string Cookie { get; set; } = "";
         public int State { get; set; }
     }
